FIX: 게시글 리스트 로딩 버그

+++ src/main/java/bssm/bsm/domain/board/post/presentation/dto/request/GetPostListRequest.java
@@ -21,5 +21,5 @@ public class GetPostListRequest {
     @NotBlank
     private String categoryId;
 
-    private int startPostId;
+    private long startPostId;
 }
